save old progress to prepare for real UI
I'm getting back my sense of smell as the symptoms clear completely and
with it liveliness. The smells of flowers make me wonder I'm only still
at the computer to write. It's a cold but very warm night, the lights of
which maybe I've never seen so fuzzy. I should be in bed.

But finally, it's play again, I'm at ease and confident this or other
will pull me out, I won't crash and burn. The sudden fear of giving up
the beautiful things I received this year was cold showers these days.

+ 180721-2000*