unveil functional n-task session interface proto
This is the conclusion of the first dips, and the code is underway
with a new iteration, which will include saving the session that did
not make the cut logistically now. It was the last thing to complete.

I want back the confidence and bliss from last log message, it truly
was sometthing. But I'm still well-rested and officially on a sort of
inverse vacation or at least a vacation of the body (which gets pain
cramped up all day on a mini-stool hunched over crap Samsung chiclet
plastic build.) The discomfort arising is from the going getting more
tough than a weekend's walk in the park, and plans boiling in my head
needing to clear out for good.

I'm also on a severe multiday calorie restriction, in part foolishly
to get some days' runway but also to adapt to the unfitting scarcity
of the weeks to come. I wasn't planning on starting back ballet class
in Copacabana daily, I was going for even more vacation, but I need
to get a good strong routine now feeding, stretching and strengthening,
otherwise I won't make it to August in shape to bear the load of every
little thing again all at once. I was maybe going back 11:30 today but
ended up sleeping till 11:00 and took the day off to wash my hair, shop
for groceries then cook in bulk. In between I'll be sending emails in
bulk too. If I don't do the chief of it on a Monday my perfectionist
aura gets dim. These chores pressure me further into coding away at
the interface prototype in complete abandon, when I should be out to
buy veggies, fruits and bottled water.