implement session saving and autoplay
There's a strong go get power surfacing now, I've played my guitar and
now going to the market. I'll start reading The Making of Karateka to
instill a habit and get going an hour daily, for inspiration. It's easy
reading and gives me inspiration for these words. Harder reading will be
taking back a chinese alchemy book, not a 100% thing, and maybe next take
Atlas Shrugged again, never finished, same for The Greatest Salesman in the
World, both easy as shallow or silly, but provide mood and counterpoint to
this necessary phase of hurdle.

I just broke up my email block by writing an attentive reply to the oldest
and most interesting conversation.