detour session utils into `soci.py` for bit of crm
My head is spinning and I'm constantly confused, the pirouettes and
the heavy influx of events make it hard to stabilize. Frequent now
are also flashes of a clear mind with program dictation. The will to
make it happen focuses the energies.

Things weren't going well with saving audio track sessions or with that
former great bout of inspired coding, so this is a new bout focused on
the hustle aspects, with session saving code repurposed, for now, to
`soci` as the module to manage relationships. Most of the work was done
yesterday I think.

Yesterday was amazing; my mind was young until she grew, it was hard to
get by in the day and spent all day at home, trying to relax. Seeing the
words generated by tnt cleanly on the bright screen under yesterday's and
today's magnificent day light makes me so happy it's hard to put into words.
It's as if I'd always deserved seeing out this expression which I have been
latching on to for seven years, a bit like being in love which is also the
case. Been feeling solar and like python is new and fun again (?). Python
flows from me now. The first months of trying to code again were awful.

But the line breaking heuristics for formatting are outrageous. It makes
the most consistent sense to break in the middle of function parameter
invocation lists even when that is radically midphrase, just to be on
the safe side of caution.

Yesterday I gave notice I'm gonna leave the apartment in one month, and on
the day before I got back on classes in Copacabana (only going again
tomorrow.) No longer going to do a new voice acting course Saturdays at
11am so have those mornings for class too. Got expelled from the group
where they announce the courses.

Everything going slowly. Very happy about this work. Very late with emails.

23 22:48 180723-210* 1700*
 + 19:23 2