stab list build mgmt interface w/ last msg preview
Actually first time I listen entirely or to any of second half of
Transformer. Great surprise and not the rosy first of coding night.

I feel a connection to the timbres of Parsons and Reed that I had
never felt before, it's along with other feelings for other people
tonight.

23:18 | Transformer (1972)
      | Let Love In (1994)
10:51 | | Tenores (José Carreras, Plácido Domingo, Luciano Pavarotti)
- - - | | 180723-220*